requirements.txt
Pillow
python-resize-image
